Fixed duplicate comment at start of group.

+++ go/cmd/format-netspoc/main.go
@@ -1,9 +1,11 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
 	"github.com/hknutzen/Netspoc/go/pkg/abort"
 	"github.com/hknutzen/Netspoc/go/pkg/conf"
+	"github.com/hknutzen/Netspoc/go/pkg/diag"
 	"github.com/hknutzen/Netspoc/go/pkg/filetree"
 	"github.com/hknutzen/Netspoc/go/pkg/parser"
 	"github.com/hknutzen/Netspoc/go/pkg/printer"
@@ -11,13 +13,17 @@ import (
 )
 
 func processInput(input *filetree.Context) {
-	bytes := []byte(input.Data)
+	source := []byte(input.Data)
 	path := input.Path
-	list := parser.ParseFile(bytes, path)
+	list := parser.ParseFile(source, path)
 	for _, n := range list {
 		n.Order()
 	}
-	copy := printer.File(list, bytes)
+	copy := printer.File(list, source)
+	if bytes.Compare(source, copy) == 0 {
+		return
+	}
+	diag.Info("Changed %s", path)
 	err := os.Remove(path)
 	if err != nil {
 		abort.Msg("Can't remove %s: %s", path, err)
